Added styling for the dashboard page and moved to its own folder to highlight difference compared to the regular layouts.
