+++ react-ui/src/components/Common/ColorSchema.scss
@@ -1,0 +1,9 @@
+$navbar-text-color: white;
+
+$student-color: #53afc6;
+$student-hover-color: #a1d9e7;
+$student-active-color: #2bb4d6;
+
+$worker-color: #c6538c;
+$worker-hover-color: #dd89b3;
+$worker-active-color: #d31976;
