+++ react-ui/src/components/Dashboard/Dashboard.scss
@@ -1,0 +1,73 @@
+@import "../Common/ColorSchema.scss";
+  
+.worker {
+  background-color: $worker-color;
+  height: 70px;
+  padding: 5px;
+
+  /* Change the color of links on hover */
+  li:hover, button {
+    background-color: $worker-hover-color;
+  }
+
+  /* Add a color to the active/current link */
+  li.active, button {
+    background-color: $worker-active-color;
+  }
+}
+
+.student {
+  background-color: $student-color;
+  height: 70px;
+  padding: 5px;
+
+  /* Change the color of links on hover */
+  li:hover, button {
+  background-color: $student-hover-color;
+  }
+
+  /* Add a color to the active/current link */
+  li.active, button {
+    background-color: $student-active-color;
+  }
+}
+
+/* Create two unequal columns that floats next to each other */
+.column {
+  float: left;
+}
+
+.left {
+  width: 25%;
+}
+
+.right {
+  width: 75%;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+/* Style the links inside the navigation bar */
+.topnav {
+  list-style-type: none;  
+
+  li {
+    float: left;
+    text-align: center;
+    padding: 14px 16px;
+    font-size: 17px;
+  }
+
+  a:link, a:visited, button {
+    color: $navbar-text-color;
+    text-decoration: none;
+    display: inline-block;
+  }
+
+  button {
+    border: none;
+    background-color: inherit;
+    font-size: 17px;
+  }
+}
