+++ react-ui/src/components/SignUpForm.js
@@ -3,7 +3,7 @@ import { Container, Row, Col } from 'react-grid-system';
 import { Redirect, Route, withRouter } from "react-router-dom";
 import Title from "./Title"
 import "./SignupForm.css"
-import dashboard from "./Layouts/Dashboard"
+// import dashboard from "./Layouts/Dashboard"
 import { Link } from 'react-router-dom';
 
 //This class is used to create a Sign-up component where users enter their name, email and password to create an account.
